requirements.txt
dnspython==2.3.0
tqdm==4.65.0
